+++ src/grasp.cpp
@@ -31,7 +31,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <math.h>
-#include <pthread.h>
 #include <iostream>
 #include <exception>
 #include <typeinfo>
@@ -1643,204 +1642,6 @@ Grasp::findOptimalContactForces(const Matrix &wrench,
   contactWrenches.copyMatrix(matrixMultiply(D, sol.getSubMatrix(1, 0, D.cols(), 1)));
 
   return -objVal;
-}
-
-/*! Evaluate PCR quality metric as described in 'Contact and Grasp 
-    Robustness Measures: Analysis and Experiment' (1997) by 
-    Prattichizzo et al. Care must be taken to pass the show=false
-    flag when calling from multiple threads as displayContactWrenches()
-    and drawObjectWrench() are not thread safe
-*/
-double
-Grasp::evaluatePCR(const Matrix &wrench, double maxForce, 
-                   std::vector<int> states /*=empty*/, bool show /*=true*/)
-{
-  //use the pre-set list of contacts. This includes contacts on the palm, but
-  //not contacts with other objects or obstacles
-  std::list<Contact*> contacts;
-  contacts.insert(contacts.begin(),contactVec.begin(), contactVec.end());
-  if (contacts.empty()) return -1;
-  
-  //check if G is full rank
-  Matrix H(contactModelMatrix(contacts.size(), states));
-  Matrix R(Contact::localToWorldWrenchBlockMatrix(contacts));
-  Matrix G(matrixMultiply(graspMapMatrix(R), H.transposed()));
-  if (G.rank() < 6) return -1;
-
-  //get only the joints on chains that make contact;
-  std::list<Joint*> joints = getJointsOnContactChains();
-
-  Matrix contactWrenches(6*contacts.size(), 1);
-  double dmin = findOptimalContactForces(wrench, maxForce, contactWrenches, joints, contacts, states);
-  if (dmin!=dmin) return -1;
-  
-  Matrix GRK(KweightedGinverse(joints, contacts, states));
-  Matrix S(std::min(GRK.rows(), GRK.cols()), 1);
-  Matrix U(GRK.rows(), GRK.rows());
-  Matrix VT(GRK.cols(), GRK.cols());
-  GRK.SVD(S,U,VT);
-
-  Matrix GKGTInv(graspStiffness(joints, contacts));
-
-  /*if (show) {
-    displayContactWrenches(&contacts, contactWrenches);
-    drawObjectWrench(wrench.negative());
-  }
-  drawObjectMovement(matrixMultiply(GKGTInv, wrench));*/
-
-  return dmin/S.elem(0,0);
-}
-
-// Allows to evaluate PCR for different contact states on a 
-// thread. This makes computation of PGR much faster. The
-// argv array of pointers contains pointers to:
-// 0. Grasp *g;
-// 1. Matrix *wrench;
-// 2. double *maxForce;
-// 3. std::vector<int> *states;
-// 4. std::vector<int> *final_states;
-// 5. double *maxPCR;
-// 6. bool *doneFlag;
-// 7. states vector mutex
-// 8. result mutex
-void*
-Grasp::evaluatePCRThread(void *thread_args)
-{
-  threadArgs *args = (threadArgs*) thread_args;
-  while(true) {
-    //Lock states mutex
-    pthread_mutex_lock(args->statesMutex_pt);
-    
-    //Check if all jobs are done 
-    if (args->doneFlag) {
-      pthread_mutex_unlock(args->statesMutex_pt);
-      pthread_exit(NULL);
-    }
-
-    //Set states for next job and increment
-    std::vector<int> states(args->states);
-    int counter;
-    do {
-      counter = 0;
-      (args->states)[0]++;
-      for (int j=0; j<args->states.size(); j++) {
-        if ((args->states)[j] == 3) {
-          (args->states)[j] = 0;
-          (args->states)[j+1] += 1;
-        }
-        if ((args->states)[j] == 0) counter += 3;
-        if ((args->states)[j] == 1) counter++;
-      }
-      //Check if we have exhausted the queue
-      args->doneFlag = true;
-      std::vector<int>::iterator it;
-      for (it=args->states.begin(); it!=args->states.end(); it++) {
-        if ((*it) != 2) {
-          args->doneFlag = false;
-          break;
-        }
-      }
-      if (args->doneFlag) {
-        pthread_mutex_unlock(args->statesMutex_pt);
-        pthread_exit(NULL);
-      }
-    } while (counter < 6);
-    pthread_mutex_unlock(args->statesMutex_pt);
-
-    //Display contact states
-    std::cerr << "Contact states: ";
-    for (std::vector<int>::iterator it=states.begin(); it!=states.end(); it++)
-      std::cerr << *it;
-    std::cerr << std::endl;
-
-    double PCR = args->g->evaluatePCR(*(args->wrench), args->maxForce, states, false);
-
-    //Lock result mutex and check if we have improved on PGR
-    pthread_mutex_lock(args->resultMutex_pt);
-    if (PCR > args->maxPCR) {
-      args->maxPCR = PCR;
-      args->finalStates = states;
-    } 
-    pthread_mutex_unlock(args->resultMutex_pt);
-  }
-}
-
-/*! Evaluate PGR quality metric as described in 'Contact and Grasp 
-    Robustness Measures: Analysis and Experiment' (1997) by 
-    Prattichizzo et al.
-*/
-double
-Grasp::evaluatePGR(Matrix &wrench, double maxForce, int maxContacts)
-{
-  //use the pre-set list of contacts. This includes contacts on the palm, but
-  //not contacts with other objects or obstacles
-  std::list<Contact*> contacts;
-  contacts.insert(contacts.begin(),contactVec.begin(), contactVec.end());
-  //if there are no contacts we are done
-  if (contacts.empty()) {
-    DBGA("No contacts");
-    return -1;
-  }
-  //if there are too many contacts, the computational cost is too high
-  if (contacts.size() > maxContacts) {
-    DBGA("Too many contacts");
-    return -1;
-  }
-
-  //Array of threads for parallelization of PGR computation
-  int numThreads = sysconf(_SC_NPROCESSORS_ONLN);
-  pthread_t threads[numThreads];
-  pthread_attr_t attr;
-  pthread_attr_init(&attr);
-  pthread_attr_setdetachstate(&attr, PTHREAD_CREATE_JOINABLE);
-  pthread_mutex_t statesMutex;
-  pthread_mutex_t resultMutex;
-  pthread_mutex_init(&statesMutex, NULL);
-  pthread_mutex_init(&resultMutex, NULL);
-
-  //Arguments for threads
-  threadArgs args;
-  args.g = this;
-  args.wrench = &wrench;
-  args.maxForce = maxForce;
-  args.states = std::vector<int>(contacts.size(), 0);
-  args.maxPCR = -1;
-  args.doneFlag = false;
-  args.statesMutex_pt = &statesMutex;
-  args.resultMutex_pt = &resultMutex;
-
-  for (int i=0; i<numThreads; i++) {
-    if (pthread_create(&threads[i], &attr, evaluatePCRThread, (void*)&args)) {
-      DBGA("Failed to create thread");
-      exit(1);
-    }
-  }
-  for (int i=0; i<numThreads; i++) {
-    void *status = NULL;
-    pthread_join(threads[i], &status);
-    if (status) {
-      DBGA("Failed to join tread;");
-      exit(1);
-    }
-  }
-
-  //Clean-up
-  pthread_attr_destroy(&attr);
-  pthread_mutex_destroy(&statesMutex);
-  pthread_mutex_destroy(&resultMutex);
-
-  if (args.maxPCR == -1) {
-    DBGA("No stable grasp found");
-    return -1;
-  }
-  double PGR = evaluatePCR(wrench, maxForce, args.finalStates);
-  assert(PGR - args.maxPCR < Matrix::EPS);
-  std::cerr << "optimal contact states: ";
-  for (std::vector<int>::iterator it=args.finalStates.begin(); it!=args.finalStates.end(); it++)
-    std::cerr << *it;
-  std::cerr << std::endl;
-
-  return PGR;
 }
 
 /*! This is a helper function for grasp force optimization routines. Given a
