+++ ui/graspCaptureDlg.cpp
@@ -31,6 +31,8 @@
 #include "robot.h"
 #include "grasp.h"
 #include "quality.h"
+#include "qualEpsilon.h"
+#include "qualVolume.h"
 #include "EGPlanner/searchState.h"
 #include "contact/contact.h"
 #include "quality.h"
